[#49] feat: viewport export 및 SidebarContent/RightRailContent 추출
* app/layout.tsx에 viewport export 추가 (themeColor, viewportFit:cover
  로 iOS safe-area 지원, maximumScale:5로 WCAG 2.5.3 준수)
* SidebarContent 컴포넌트 신규 — 데스크톱/모바일 공유용. 상단에
  Search & jump 버튼(CommandPalette 진입) 추가, 모바일 터치 타겟을
  위한 py-2.5 md:py-1 적용, onNavigate prop으로 drawer 닫힘 제어
* RightRailContent 컴포넌트 신규 — CommitGraph/Latest posts/System
  블록을 데스크톱/모바일 공유
* Sidebar.tsx / RightRail.tsx를 각각 hidden md:flex / hidden xl:flex
  래퍼로 축소
